added static_cast to c08

+++ C08/ex01/inc/Span.tpp
@@ -4,7 +4,7 @@
 template <typename InputIterator>
 void Span::addNumbers(InputIterator first, InputIterator last)
 {
-	size_t numToAdd = std::distance(first, last);
+	size_t numToAdd = static_cast<size_t>(std::distance(first, last));
 	if (_valuesTab.size() + numToAdd > _maxSize)
 		throw std::out_of_range("Error: adding values would exceed maximum size");
 	if (_valuesTab.size() + numToAdd == 0)
